refactor: move interface in right package
