+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/rest/api/service/v4/impl/validation/FlowValidationServiceImpl.java
@@ -19,8 +19,8 @@ import io.gravitee.definition.model.v4.flow.Flow;
 import io.gravitee.definition.model.v4.flow.selector.Selector;
 import io.gravitee.rest.api.service.PolicyService;
 import io.gravitee.rest.api.service.impl.TransactionalService;
-import io.gravitee.rest.api.service.v4.FlowValidationService;
 import io.gravitee.rest.api.service.v4.exception.FlowSelectorsDuplicatedException;
+import io.gravitee.rest.api.service.v4.validation.FlowValidationService;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
