+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/test/java/io/gravitee/rest/api/service/v4/impl/validation/FlowValidationServiceImplTest.java
@@ -28,8 +28,8 @@ import io.gravitee.definition.model.v4.flow.selector.SelectorHttp;
 import io.gravitee.definition.model.v4.flow.step.Step;
 import io.gravitee.rest.api.service.PolicyService;
 import io.gravitee.rest.api.service.exceptions.InvalidDataException;
-import io.gravitee.rest.api.service.v4.FlowValidationService;
 import io.gravitee.rest.api.service.v4.exception.FlowSelectorsDuplicatedException;
+import io.gravitee.rest.api.service.v4.validation.FlowValidationService;
 import java.util.List;
 import org.junit.Before;
 import org.junit.Test;
